Rework visibility under doors and handle apport
It turns out that making items under doors invisible was too strong;
this commit adds the distinction between items that are both visible and
accessible for manipulation, and items that are visible.  Apport,
animate dead (stack_iterator in general) use the former.

This commit also cleans up apport in deep water/lava to use
visible_igrd, so that species that can see items in deep water can
apport them.  (It also removes the silly "density" message when
apporting in deep water.)